Only authenticated users can now upload, edit and delete pizzas from the Database

+++ pages/api/products/[id].js
@@ -9,7 +9,9 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+  const token = cookies.token;
 
   if (method === 'GET') {
     try {
@@ -25,6 +27,10 @@ export default async function handler(req, res) {
     }
   }
   if (method === 'PUT') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('Not authenticated');
+    }
+
     try {
       const product = await Product.findByIdAndUpdate(id);
       res.status(201).json(product);
@@ -34,6 +40,10 @@ export default async function handler(req, res) {
   }
 
   if (method === 'DELETE') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('Not authenticated');
+    }
+
     try {
       await Product.findByIdAndDelete(id);
       res.status(201).json('The product has been deleted!');
